Update movie controller, service, entity
Update the movie entity to correctly store the release date as a date object in the underlying database instead of a string. Add one to many mapping for review associated with a given movie.

Add post mapping to controller to expose a post endpoint which allows multiple movies to be added together, i.e. a list.

Add method in service to add a collection of movies to the database via the repository.